added index file pattern
